pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "discord-music-bot"
version = "0.1.0"
description = "A self-hosted Discord music bot that streams audio from YouTube, Spotify, and SoundCloud"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.0",
    "yt-dlp",
    "spotipy",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[tool.coverage.run]
source = ["bot"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
